core: collect all code relating to lowpriv raise into one file
Similar to how I already extracted `type ResourceBehavior` into its own
file instead of having it scattered in `cluster.go` and `config.go`.
