+++ internal/core/cluster.go
@@ -22,9 +22,7 @@ package core
 import (
 	"context"
 	"os"
-	"regexp"
 	"sort"
-	"strconv"
 
 	"github.com/gophercloud/gophercloud"
 	"github.com/open-policy-agent/opa/rego"
@@ -47,11 +45,8 @@ type Cluster struct {
 	CapacityPlugins   map[string]CapacityPlugin
 	Authoritative     bool
 	QuotaConstraints  *QuotaConstraintSet
-	LowPrivilegeRaise struct {
-		LimitsForDomains  map[string]map[string]LowPrivilegeRaiseLimit
-		LimitsForProjects map[string]map[string]LowPrivilegeRaiseLimit
-	}
-	OPA struct {
+	LowPrivilegeRaise LowPrivilegeRaiseLimitSet
+	OPA               struct {
 		ProjectQuotaQuery *rego.PreparedEvalQuery
 		DomainQuotaQuery  *rego.PreparedEvalQuery
 	}
@@ -207,16 +202,8 @@ func (c *Cluster) Connect(provider *gophercloud.ProviderClient, eo gophercloud.E
 	}
 
 	//parse low-privilege raise limits
-	c.LowPrivilegeRaise.LimitsForDomains, suberrs = c.parseLowPrivilegeRaiseLimits(
-		c.Config.LowPrivilegeRaise.Limits.ForDomains, "domain")
-	for _, err := range suberrs {
-		errs.Addf("while parsing low-privilege raise limit for domains: %w", err)
-	}
-	c.LowPrivilegeRaise.LimitsForProjects, suberrs = c.parseLowPrivilegeRaiseLimits(
-		c.Config.LowPrivilegeRaise.Limits.ForProjects, "project")
-	for _, err := range suberrs {
-		errs.Addf("while parsing low-privilege raise limit for projects: %w", err)
-	}
+	c.LowPrivilegeRaise, suberrs = c.Config.LowPrivilegeRaise.parse(c.QuotaPlugins)
+	errs.Append(suberrs)
 
 	//validate scaling relations
 	for _, behavior := range c.Config.ResourceBehaviors {
@@ -228,71 +215,6 @@ func (c *Cluster) Connect(provider *gophercloud.ProviderClient, eo gophercloud.E
 	}
 
 	return nil
-}
-
-var (
-	percentOfClusterRx              = regexp.MustCompile(`^([0-9.]+)\s*% of cluster capacity$`)
-	untilPercentOfClusterAssignedRx = regexp.MustCompile(`^until ([0-9.]+)\s*% of cluster capacity is assigned$`)
-)
-
-func (c Cluster) parseLowPrivilegeRaiseLimits(inputs map[string]map[string]string, scopeType string) (result map[string]map[string]LowPrivilegeRaiseLimit, errs ErrorSet) {
-	result = make(map[string]map[string]LowPrivilegeRaiseLimit)
-	for srvType, quotaPlugin := range c.QuotaPlugins {
-		result[srvType] = make(map[string]LowPrivilegeRaiseLimit)
-		for _, res := range quotaPlugin.Resources() {
-			limit, exists := inputs[srvType][res.Name]
-			if !exists {
-				continue
-			}
-
-			match := percentOfClusterRx.FindStringSubmatch(limit)
-			if match != nil {
-				percent, err := strconv.ParseFloat(match[1], 64)
-				if err != nil {
-					errs.Add(err)
-					continue
-				}
-				if percent < 0 || percent > 100 {
-					errs.Addf("value out of range: %s%%", match[1])
-					continue
-				}
-				result[srvType][res.Name] = LowPrivilegeRaiseLimit{
-					PercentOfClusterCapacity: percent,
-				}
-				continue
-			}
-
-			//the "until X% of cluster capacity is assigned" syntax is only allowed for domains
-			if scopeType == "domain" {
-				match := untilPercentOfClusterAssignedRx.FindStringSubmatch(limit)
-				if match != nil {
-					percent, err := strconv.ParseFloat(match[1], 64)
-					if err != nil {
-						errs.Add(err)
-						continue
-					}
-					if percent < 0 || percent > 100 {
-						errs.Addf("value out of range: %s%%", match[1])
-						continue
-					}
-					result[srvType][res.Name] = LowPrivilegeRaiseLimit{
-						UntilPercentOfClusterCapacityAssigned: percent,
-					}
-					continue
-				}
-			}
-
-			rawValue, err := res.Unit.Parse(limit)
-			if err != nil {
-				errs.Add(err)
-				continue
-			}
-			result[srvType][res.Name] = LowPrivilegeRaiseLimit{
-				AbsoluteValue: rawValue,
-			}
-		}
-	}
-	return result, nil
 }
 
 // ServiceTypesInAlphabeticalOrder can be used when service types need to be
